+++ code/modules/mob/new_player/new_player.dm
@@ -37,6 +37,8 @@
 
 
 /mob/new_player/proc/check_playerpolls()
+	if(!WHITELIST_CHECK(client))
+		return FALSE
 	var/output
 	if (SSdbcore.Connect())
 		var/isadmin = FALSE
@@ -74,6 +76,8 @@
 	. = ..()
 
 	if(!SSticker)
+		return
+	if(!WHITELIST_CHECK(client))
 		return
 
 	if(SSticker.current_state == GAME_STATE_PREGAME)
@@ -143,6 +147,9 @@
 			view_som()
 
 		if("SelectedJob")
+			if(!WHITELIST_CHECK(client))
+				WHITELIST_MESSAGE(client)
+				return
 			if(!SSticker)
 				return
 			if(!GLOB.enter_allowed)
@@ -207,6 +214,9 @@
 
 
 /mob/new_player/proc/late_choices()
+	if(!WHITELIST_CHECK(client))
+		WHITELIST_MESSAGE(client)
+		return
 	var/list/dat = list("<div class='notice'>Round Duration: [DisplayTimeText(world.time - SSticker.round_start_time)]</div>")
 	if(!GLOB.enter_allowed)
 		dat += "<div class='notice red'>You may no longer join the round.</div><br>"
@@ -374,6 +384,9 @@
 
 /mob/new_player/proc/transfer_character()
 	. = new_character
+	if(!WHITELIST_CHECK(client))
+		WHITELIST_MESSAGE(client)
+		return
 	if(.)
 		mind.transfer_to(new_character, TRUE) //Manually transfer the key to log them in
 		qdel(src)
@@ -399,13 +412,16 @@
 	return TRUE
 
 /mob/new_player/proc/try_to_observe()
+	if(!WHITELIST_CHECK(client))
+		WHITELIST_MESSAGE(client)
+		return
 	if(!SSticker || SSticker.current_state == GAME_STATE_STARTUP)
 		to_chat(src, span_warning("The game is still setting up, please try again later."))
 		return
 	if(client?.observe_used)
 		to_chat(src,  span_warning("You seen enough, time to play."))
 		return FALSE
-	if(!check_other_rights(client, R_ADMIN, FALSE))
+	if(!check_other_rights(client, R_ADMIN, FALSE) && SSticker.mode.round_type_flags2 & MODE_2_NO_GHOSTS)
 		log_game("[key_name(src)] failed to join as a ghost due to the observe disable.")
 		to_chat(src, span_boldannounce("Observing is currently disabled.  Please do not get around this by joining just to ghost."))
 		spawn()
@@ -474,6 +490,10 @@
 		return
 	ready = !ready
 	if(ready)
+		if(!WHITELIST_CHECK(client))
+			WHITELIST_MESSAGE(client)
+			ready = FALSE
+			return
 		GLOB.ready_players += src
 	else
 		GLOB.ready_players -= src
@@ -481,6 +501,9 @@
 
 ///Attempts to latejoin the player
 /mob/new_player/proc/attempt_late_join(queue_override = FALSE)
+	if(!WHITELIST_CHECK(client))
+		WHITELIST_MESSAGE(client)
+		return
 	if(!SSticker?.mode || SSticker.current_state != GAME_STATE_PLAYING)
 		to_chat(src, span_warning("The round is either not ready, or has already finished."))
 		return
@@ -516,16 +539,22 @@
 
 
 /mob/new_player/proc/take_ssd_mob()
+	if(!WHITELIST_CHECK(client))
+		WHITELIST_MESSAGE(client)
 	if((src.key in GLOB.key_to_time_of_death) && (GLOB.key_to_time_of_death[src.key] + TIME_BEFORE_TAKING_BODY > world.time))
 		to_chat(src, span_warning("You died too recently to be able to take a new mob."))
 		return
 
-
-	var/list/mob/living/free_ssd_mobs = GLOB.offered_mob_list
+	var/list/mob/living/free_ssd_mobs = list()
+	var/list/mob/living/mobs_to_check = GLOB.ssd_living_mobs
+	mobs_to_check += GLOB.offered_mob_list
 	if(GLOB.ssd_posses_allowed)
-		for(var/mob/living/ssd_mob AS in GLOB.ssd_living_mobs)
-			if(is_centcom_level(ssd_mob.z) || ishuman(ssd_mob) || ssd_mob.afk_status == MOB_RECENTLY_DISCONNECTED)
+		for(var/mob/living/ssd_mob AS in mobs_to_check)
+			if(is_centcom_level(ssd_mob.z) || ssd_mob.afk_status == MOB_RECENTLY_DISCONNECTED)
 				continue
+			if(ishuman(ssd_mob))
+				if(length(ssd_mob.ckey_history) && !(key in ssd_mob.ckey_history)) //can only take your own human characters' control unless they are empty mobs from the get go.
+					continue
 			free_ssd_mobs += ssd_mob
 
 	if(!length(free_ssd_mobs))
@@ -546,7 +575,6 @@
 		if(ssd_xeno.tier != XENO_TIER_MINION && XENODEATHTIME_CHECK(src))
 			XENODEATHTIME_MESSAGE(src)
 			return
-
 	if(HAS_TRAIT(new_mob, TRAIT_POSSESSING))
 		to_chat(src, span_warning("That mob is currently possessing a different mob."))
 		return FALSE
@@ -583,5 +611,5 @@
 	var/mob/living/carbon/human/H = new_mob
 	var/datum/job/j = H.job
 	var/datum/outfit/job/o = j.outfit
-	H.on_transformation()
+	//H.on_transformation()
 	o.handle_id(H)
